refactor, move generic cidr code to `lib/cidr.go`
